requirements.txt
annotated-types==0.7.0
anyio==3.7.1
babel==2.17.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
courlan==1.3.2
cssselect==1.3.0
cssutils==2.11.1
dateparser==1.2.2
distro==1.9.0
feedparser==6.0.10
greenlet==3.2.4
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
more-itertools==10.7.0
openai==1.99.9
premailer==3.10.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
regex==2025.7.34
requests==2.32.4
setuptools==78.1.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
tld==0.13.1
tqdm==4.67.1
trafilatura==1.6.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
yagmail==0.15.293
tenacity==8.2.3
